+++ src/communication_structures.h
@@ -77,6 +77,8 @@ const int SOILEBCOM_k_down = 13;
 const int SOILEBCOM_tempch = 14;
 List communicationSoilEnergyBalance(int nlayers);
 
+NumericVector communicationFireHazard();
+
 List basicTranspirationCommunicationOutput(int numCohorts, int nlayers);
 List advancedTranspirationCommunicationOutput(int numCohorts, int nlayers, int ncanlayers, int ntimesteps);
 List copyBasicTranspirationOutput(List btc, List x);